+++ Sources/KeyboardKit/_Deprecated/DeviceSpecificInputSetProvider.swift
@@ -15,33 +15,38 @@ import Foundation
  For now, this provider will return the iPhone input set for
  any device that is not explicitly iPad.
  */
+@available(*, deprecated, message: "DeviceSpecificInputSetProvider is no longer needed")
 public protocol DeviceSpecificInputSetProvider: InputSetProvider {}
 
+@available(*, deprecated, message: "DeviceSpecificInputSetProvider is no longer needed")
 public extension DeviceSpecificInputSetProvider {
-    
+
     /**
      Create an input row from phone and pad-specific strings.
      */
+    @available(*, deprecated, message: "Use InputSetRow initializer instead")
     func row(
         phone: String,
         pad: String,
         deviceType: DeviceType = .current) -> InputSetRow {
         InputSetRow(deviceType == .pad ? pad.chars : phone.chars)
     }
-    
+
     /**
      Create an input row from phone and pad-specific arrays.
      */
+    @available(*, deprecated, message: "Use InputSetRow initializer instead")
     func row(
         phone: [String],
         pad: [String],
         deviceType: DeviceType = .current) -> InputSetRow {
         InputSetRow(deviceType == .pad ? pad : phone)
     }
-    
+
     /**
      Create an input row from phone and pad-specific strings.
      */
+    @available(*, deprecated, message: "Use InputSetRow initializer instead")
     func row(
         phoneLowercased: String,
         phoneUppercased: String,
@@ -52,10 +57,11 @@ public extension DeviceSpecificInputSetProvider {
             lowercased: deviceType == .pad ? padLowercased.chars : phoneLowercased.chars,
             uppercased: deviceType == .pad ? padUppercased.chars : phoneUppercased.chars)
     }
-    
+
     /**
      Create an input row from phone and pad-specific arrays.
      */
+    @available(*, deprecated, message: "Use InputSetRow initializer instead")
     func row(
         phoneLowercased: [String],
         phoneUppercased: [String],